run code cleanup and refactor using rider default config

+++ social-campus-server/Infrastructure/Repositories/PublicationLikeRepositories.cs
@@ -5,38 +5,35 @@ using Domain.Models.UserModel;
 using Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
 
-namespace Infrastructure.Repositories
+namespace Infrastructure.Repositories;
+
+public class PublicationLikeRepositories(
+    ApplicationDbContext context) : IPublicationLikeRepositories
 {
-    public class PublicationLikeRepositories(
-        ApplicationDbContext context) : IPublicationLikeRepositories
+    public async Task AddAsync(UserId userId, PublicationId publicationId)
     {
-        public async Task AddAsync(UserId userId, PublicationId publicationId)
-        {
-            PublicationLike newPublicationLike = new(userId, publicationId);
+        PublicationLike newPublicationLike = new(userId, publicationId);
 
-            await context.PublicationLikes.AddAsync(newPublicationLike);
-        }
+        await context.PublicationLikes.AddAsync(newPublicationLike);
+    }
 
-        public async Task DeleteAsync(UserId userId, PublicationId publicationId)
-        {
-            PublicationLike? publicationLike = await context.PublicationLikes
-                .FirstOrDefaultAsync(pl => pl.PublicationId == publicationId && pl.UserId == userId);
-            if (publicationLike != null)
-            {
-                context.PublicationLikes.Remove(publicationLike);
-            }
-        }
+    public async Task DeleteAsync(UserId userId, PublicationId publicationId)
+    {
+        var publicationLike = await context.PublicationLikes
+            .FirstOrDefaultAsync(pl => pl.PublicationId == publicationId && pl.UserId == userId);
+        if (publicationLike != null) context.PublicationLikes.Remove(publicationLike);
+    }
 
-        public async Task<IReadOnlyList<PublicationLike>> GetPublicationLikesListByPublicationIdAsync(PublicationId publicationId)
-        {
-            return await context.PublicationLikes
-                .Where(pl => pl.PublicationId == publicationId)
-                .ToListAsync() as IReadOnlyList<PublicationLike>;
-        }
+    public async Task<IReadOnlyList<PublicationLike>> GetPublicationLikesListByPublicationIdAsync(
+        PublicationId publicationId)
+    {
+        return await context.PublicationLikes
+            .Where(pl => pl.PublicationId == publicationId)
+            .ToListAsync();
+    }
 
-        public async Task<bool> IsLike(UserId userId, PublicationId publicationId)
-        {
-            return await context.PublicationLikes.AnyAsync(pl => pl.PublicationId == publicationId && pl.UserId == userId);
-        }
+    public async Task<bool> IsLike(UserId userId, PublicationId publicationId)
+    {
+        return await context.PublicationLikes.AnyAsync(pl => pl.PublicationId == publicationId && pl.UserId == userId);
     }
 }
